+++ src/components/player-profile-picture/player-profile-picture.component.js
@@ -1,7 +1,14 @@
 module.exports = {
     onInput(input) {
+        let profilePicUrl = input.profile_pic_url;
+        let playerId = input.player_id;
+        if (input.player) {
+            playerId = input.player.id;
+            profilePicUrl = input.player.profile_pic_url;
+        }
         this.state = {
-            profile_pic_url: input.player ? input.player.profile_pic_url : undefined
+            profile_pic_url: profilePicUrl,
+            player_id: playerId
         }
     }
 }